services/scene_recognition/requirements.txt
open-clip-torch
torch
transformers
pillow
fastapi
uvicorn
python-multipart 